+++ app/models/__init__.py
@@ -1,7 +1,0 @@
-from app.models.Route import Route
-from app.models.Journey import Journey
-from app.models.Route import Stop
-
-
-
-